requirements.txt
pydantic[dotenv]~=1.8.2
aioredis[hiredis]~=2.0.0
python-dotenv~=0.19.0
dnspython~=2.1.0
loguru~=0.5.3
aiorun~=2021.8.1
aiohttp[speedups]~=3.7.4